refactor: ディレクトリ構成の見直し - ref: [Flutter Project Structure: Feature-first or Layer-first?](https://codewithandrea.com/articles/flutter-project-structure/#feature-first-layers-inside-features)
