+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_training/ui/app.dart';
+import 'package:flutter_training/app.dart';
 
 void main() {
   runApp(const App());
